Make it so LinesOfText roundtrips to text
We were not tracking whether the last line contained or not a
new line, so it was impossible it roundtrip. We now keep the
newline on every line, except on the last one where it is optional.

We also add an invariant that says that the vector of lines must
be non-empty (since we would otherwise have two representations for
"", which is annoying).